Reworked quite a bit of the splitter's code to now use the new memory layer. Good chance a lot of stuff is now broken, it's definitly untested.
