+++ src/DarkSoulsMemory/DarkSoulsMemory.csproj
@@ -58,12 +58,11 @@
     <Compile Include="Internal\Extensions.cs" />
     <Compile Include="ForcedAnimation.cs" />
     <Compile Include="Internal\IDarkSouls.cs" />
-    <Compile Include="Internal\Item.cs" />
-    <Compile Include="Internal\ItemPrompt.cs" />
+    <Compile Include="Item.cs" />
+    <Compile Include="ItemPrompt.cs" />
     <Compile Include="Internal\Kernel32.cs" />
     <Compile Include="MenuPrompt.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Internal\Zone.cs" />
     <Compile Include="ZoneType.cs" />
   </ItemGroup>
   <ItemGroup>
