+++ src/DarkSoulsMemory/Item.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DarkSoulsMemory.Internal
+namespace DarkSoulsMemory
 {
     public class Item
     {
