+++ src/LiveSplit.DarkSouls/Data/SplitTypes.cs
@@ -17,7 +17,6 @@ namespace LiveSplit.DarkSouls.Data
 		Manual,
 		Quitout,
 		Zone,
-		Remastered,
 
 		// This value allows splits to be created and always exist for a split line, even if no type/data has been
 		// selected.
