+++ src/LiveSplit.DarkSouls/LiveSplit.DarkSouls.csproj
@@ -21,6 +21,7 @@
     <DefineConstants>DEBUG;TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
+    <PlatformTarget>x64</PlatformTarget>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
     <DebugType>pdbonly</DebugType>
@@ -42,6 +43,10 @@
     <Reference Include="System.Core" />
     <Reference Include="System.Data" />
     <Reference Include="System.Drawing" />
+    <Reference Include="System.Numerics" />
+    <Reference Include="System.Numerics.Vectors, Version=4.1.4.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL">
+      <HintPath>..\..\packages\System.Numerics.Vectors.4.5.0\lib\net46\System.Numerics.Vectors.dll</HintPath>
+    </Reference>
     <Reference Include="System.Windows.Forms" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="Microsoft.CSharp" />
@@ -80,7 +85,6 @@
     <Compile Include="Data\ItemState.cs" />
     <Compile Include="Data\ItemTracker.cs" />
     <Compile Include="Data\TrackerBase.cs" />
-    <Compile Include="Data\Vector3.cs" />
     <Compile Include="Data\Zone.cs" />
     <Compile Include="Kernel32.cs" />
     <Compile Include="Memory\Flags.cs" />
@@ -96,7 +100,6 @@
     <Compile Include="Memory\MemoryTools.cs" />
     <Compile Include="Memory\SoulsPointers.cs" />
     <Compile Include="Memory\SoulsMemory.cs" />
-    <Compile Include="Memory\SoulsRemastered.cs" />
     <Compile Include="Resources.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
@@ -125,6 +128,7 @@
       <DependentUpon>SoulsForm.cs</DependentUpon>
     </EmbeddedResource>
     <None Include="app.config" />
+    <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Controls\SoulsMasterControl.resx">
