+++ src/LiveSplit.DarkSouls/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net461" />
+</packages>
